tests: suppress mgm api timeout error in sotrcon (#11428)
## Problem
Since
https://github.com/neondatabase/neon/commit/084a2d6180d2906ed33558ba0aeffe9cc8b05bcb
the timeout in `with_client_retries` is implemented via `tokio::timeout`
instead of `reqwest::ClientBuilder::timeout` (because we reuse the
client). It changed the error representation if the timeout is exceeded.
Such errors were suppressed in `allowed_errors.py`, but old regexps do
not match the new error.
Discussion:
https://neondb.slack.com/archives/C033RQ5SPDH/p1743533184736319

## Summary of changes
- Add new `Timeout` error to `allowed_errors.py`